Whew. The weekend.
I've reinstated the database on Amazon, using SQL now, instead of derby
on dropbox. That shit got mad corrupted. 

Basically now, Location add and find pages work.

I've just been working on the Artist add page... the save method for
artist is pretty gnarly, it uses a wrapper class to hold all of the
fields that get separated out into their right entity in the
ArtistController. 

Added a metaplay.utilities.DateFormatter to format dates from Strings.
Will need to make a note to the user on the artist_add page that
notifies them of the format: YYYY-MM-DD.

Good work this weekend!