Wohooo! THE ARTIST PAGE WORKS.
A method that calls a creation method involving a transaction has to
ALSO be annotated with @Transactional... geez.

Fixed some other things like the Collection Adders and Updaters, to make
sure they were not calling methods on null objects (like for instance,
if there were no Members to begin with, and you were trying to add a
Member).