Replaced LocationService with LocationRepository.
LocationRepository implements JpaRepository, as well as
LocationRepositoryCustom, allowing me to create custom functionality in
my queries and logic. Now, I've inherited 18 methods from JpaRepository
as well. This helps eliminate boilerplate code, as well as gets rid of
the need to use EntityManagerFactory all over the place. EMF is not
thread-safe, and "generally shouldn't be injected into a shared
singleton bean" (*In fact, it's not even injected, but you're given a
proxy to the 'real' EMF).