**Fixed the Artist Add page to work with the new JPARepository methods.
For an unknown reason, when I was using the JPA/Spring Data methods such
as findAlbumByNameIsNotNull(), I would get a NoSuchElementException. It
seems something really did not like this type of IsNotNull method,
because they were all causing problems: the Member version, the Album
version, etc.

I gave up and got rid of them, and instead am just checking manually:
!=null. It's the same functionality...

**Fixed the memberRepository.splitFullName method. Gnarly. The use of
this method and the 2d array is kind of cool to loop through the members
and set their full name and stage name if the user entered them. Pretty
cool.

**Added functionality to boot a user out of viewing another user's
profile page. I need to throw an error when this happens, as that's
likely some kind of hack fuckery going on. You now get directed back to
your profile page if you try to url hack to another user's profile page.