Major new functionality: profile pictures.
Storing them in AWS S3. Pretty cool. 

Also attempted working with SSL for a while, trying to configure tomcat
with HTTPS on port 8443, but couldn't get that to work. Mostly, the
stuff is commented out so I can try again later.

Still need to test a lot of the add pages, and continue to work on
exception handling.