Input boxes now focus like they should.
Fixed the account_login page to actually display -- issue with the delayRevealed class resolved for this specific page.

Added tracks iteration to the album page, so that you can see its tracks; they link to single_ pages, too.

Changed the font color to slategray, and darkseagreen for the security info at the top of a page.

Sketched out disabling the save playlist button on click, still need to test this.

Still have a persistence issue for playlist, where the server needs to be restarted for the relationships to be seen (synchronized).