** Couple of changes: **
Added a small note to the browse_playlist.jsp page to indicate which
playlist if any you own (only appears if you're logged in).

Edited the drSecurity css that displays a note about inability to stream
music if you're anonymous.

Added links to the emails that appear on account profiles.

Changed the fonts of the body and h1 tags to be gray-scaled instead of
'slategray'.